Update manjaro_help.sh
If the list of HOOKS is different from the one chosen in the script on a system (for example, when luks encryption is used), blind execution of the script can cause Manjaro to stop booting properly because the previous options are overridden and some of the hooks are lost in the process. These changes fix that and allow the addition of the necessary hooks without overriding the previous ones.